+++ src/main/java/firok/tiths/traits/TraitPeaceEnergetic.java
@@ -1,16 +1,15 @@
 package firok.tiths.traits;
 
 import firok.tiths.common.Configs;
-import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import slimeknights.tconstruct.library.traits.AbstractTrait;
 
 import static firok.tiths.common.Keys.colorTraitPeaceEnergetic;
 import static firok.tiths.common.Keys.nameTraitPeaceEnergetic;
-import static firok.tiths.util.Predicates.canTrigger;
+import static firok.tiths.util.Predicates.canTick;
 
 // 平和能量
 public class TraitPeaceEnergetic extends AbstractTrait
@@ -21,20 +20,14 @@ public class TraitPeaceEnergetic extends AbstractTrait
 	}
 
 	@Override
-	public void afterHit(ItemStack tool, EntityLivingBase player, EntityLivingBase target, float damageDealt, boolean wasCritical, boolean wasHit)
+	public void onUpdate(ItemStack tool, World world, Entity entity, int itemSlot, boolean isSelected)
 	{
-		if(!player.world.isRemote && canTrigger(player.world, Configs.Traits.rate_peace_energetic_hit))
+		if(isSelected && !world.isRemote &&
+		   entity.motionX==0 && entity.motionY==0 && entity.motionZ==0 &&
+		   canTick(world,20,1) && entity instanceof EntityLivingBase)
 		{
-			player.heal(Configs.Traits.factor_peace_energetic_heal);
-		}
-	}
-
-	@Override
-	public void afterBlockBreak(ItemStack tool, World world, IBlockState state, BlockPos pos, EntityLivingBase player, boolean wasEffective)
-	{
-		if(!player.world.isRemote && canTrigger(player.world, Configs.Traits.rate_peace_energetic_break))
-		{
-			player.heal(Configs.Traits.factor_peace_energetic_heal);
+			EntityLivingBase enlb=(EntityLivingBase)entity;
+			enlb.heal(Configs.Traits.factor_peace_energetic_heal);
 		}
 	}
 }